add bottom Navigation

+++ app/src/main/java/com/shana/foodandgrocery/data/LocalDataSource.kt
@@ -58,7 +58,7 @@ class LocalDataSource @Inject constructor(
                 RecipeExtendedIngredientCrossRefs.add(
                     RecipeExtendedIngredientCrossRefEntity(
                         recipeId = item.id,
-                        id = id.toInt()
+                        id = id
                     )
                 )
             }
